+++ atrium-api/src/com/atproto/server/create_invite_code.rs
@@ -5,7 +5,7 @@
 pub struct Input {
     #[serde(skip_serializing_if = "Option::is_none")]
     pub for_account: Option<String>,
-    pub use_count: i32,
+    pub use_count: i64,
 }
 #[derive(serde::Serialize, serde::Deserialize, Debug, Clone, PartialEq, Eq)]
 #[serde(rename_all = "camelCase")]
